refactor: update import paths to use alias mapping
- Replace relative imports with @/ alias for better maintainability
- Update Prisma client imports to use @generated/prisma path
- Maintain consistent import structure across all modules
- Align with TypeScript path configuration changes
