+++ src/board-followers/board-followers.service.ts
@@ -1,6 +1,6 @@
+import { PrismaService } from '@/prisma/prisma.service';
+import { BoardFollower, Prisma } from '@generated/prisma';
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { BoardFollower, Prisma } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
 export class BoardFollowersService {
